TestVersion for ProfilePictures Ref #39

+++ backend/controllers/controllerProfilePicture.js
@@ -29,6 +29,7 @@ async function uploadProfilePicture(req, res) {
             const newFilename = `Picture${userID}${fileExtension}`;
 
             const filePath = path.join(uploadOptions.uploadDir, newFilename);
+            const filePathCompressed = path.join(uploadOptions.uploadDirCom, newFilename)
             const pictureURL = await database.getProfilePictureURL(userIDObj);
             console.log(filePath);
 
@@ -41,7 +42,18 @@ async function uploadProfilePicture(req, res) {
                     console.log(`Bild ${pictureURL} erfolgreich gelöscht`);
 
                 });
+            const pictureURLCompressed = await database.getProfilePictureURLCompressed(userIDObj)
+            if (pictureURLCompressed != null) {
+                fs.unlink(pictureURLCompressed, (err) => {
+                    if (err) {
+                        console.error('Fehler beim Löschen des Bildes aus dem Dateisystem:', err);
+                        return;
+                    }
+                    console.log(`Bild ${pictureURL} erfolgreich gelöscht`);
+
+                });
                 const deleteURL = await database.deleteProfilePictureURL(userIDObj);
+                const deleteURLCompressed = await database.deleteProfilePictureURLCompressed(userIDObj)
                 const uploadPicture = await database.uploadProfilePicture(userIDObj, fileExtensionParam);
                 console.log("deleteURL:" + deleteURL);
                 console.log("PictureUploadWithDelete:"+uploadPicture)
@@ -59,7 +71,7 @@ async function uploadProfilePicture(req, res) {
             });
 
             //uploadedFilename = newFilename;
-        })
+        }})
         form.on('error', () => { })
         form.on('close',async () => {
